requirements.txt
aiohttp==3.6.2
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
click==7.1.2
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.4.0
djangorestframework==3.11.1
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6
psycopg2==2.8.5
pyee==7.0.2
python-dotenv==0.14.0
pytz==2020.1
slackclient==2.8.0
slackeventsapi==2.2.1
sqlparse==0.3.1
Werkzeug==1.0.1
whitenoise==5.2.0
yarl==1.5.1
